Port DEFVAR_PER_BUFFER macro
As well as the defvar_per_buffer function. Additionally, moved the
Lisp_Fwd union and related structs into data.rs, as we're nearly ready
to completely Rustify them. Lisp_Buffer_Objfwd and Lisp_Kboard_Objfwd
now store a FieldOffset.

+++ src/lisp.h
@@ -2542,7 +2542,7 @@ struct Lisp_Objfwd
 struct Lisp_Buffer_Objfwd
   {
     enum Lisp_Fwd_Type type;	/* = Lisp_Fwd_Buffer_Obj */
-    int offset;
+    size_t offset;
     /* One of Qnil, Qintegerp, Qsymbolp, Qstringp, Qfloatp or Qnumberp.  */
     Lisp_Object predicate;
   };
@@ -2599,7 +2599,7 @@ struct Lisp_Buffer_Local_Value
 struct Lisp_Kboard_Objfwd
   {
     enum Lisp_Fwd_Type type;	/* = Lisp_Fwd_Kboard_Obj */
-    int offset;
+    size_t offset;
   };
 
 union Lisp_Fwd
